Number input improvements and brightness float values

+++ src/PixiEditor.AvaloniaUI/Models/Handlers/Tools/IBrightnessToolHandler.cs
@@ -8,5 +8,5 @@ internal interface IBrightnessToolHandler : IToolHandler
     public BrightnessMode BrightnessMode { get; }
     public bool Darken { get; }
     public MouseButton UsedWith { get; }
-    public int CorrectionFactor { get; }
+    public float CorrectionFactor { get; }
 }
